RF_015.01.02 | Support > FAQ > Visibility, сlickability, correct links redirection and dropdown > Verify 'FAQ' is clickable
