+++ tests/test_group_snake_oil/tests/test_main_page.py
@@ -13,3 +13,15 @@ class TestMainPage:
         main_page = MainPage(driver, HOME_PAGE_URL)
         main_page.open_page()
         main_page.check_clickability_of_linkedIn_icon()
+
+    def test_tc_015_01_01_verify_support_faq_is_visible(self, driver):
+        support_page = MainPage(driver, HOME_PAGE_URL)
+        support_page.open_page()
+        support_page.open_Support_dropdown()
+        support_page.check_visibility_of_FAQ_element()
+
+    def test_tc_015_01_02_verify_support_faq_is_clickable(self, driver):
+        support_page = MainPage(driver, HOME_PAGE_URL)
+        support_page.open_page()
+        support_page.open_Support_dropdown()
+        support_page.check_click_FAQ_element()
